+++ ForestManagement/functions/exchangeschema/Get-FMExchangeSchema.ps1
@@ -1,0 +1,24 @@
+﻿function Get-FMExchangeSchema
+{
+<#
+	.SYNOPSIS
+		Returns the defined Exchange Forest configuration to apply.
+	
+	.DESCRIPTION
+		Returns the defined Exchange Forest configuration to apply.
+	
+	.EXAMPLE
+		PS C:\> Get-FMExchangeSchema
+	
+		Returns the defined Exchange Forest configuration to apply.
+#>
+	[CmdletBinding()]
+	Param (
+	
+	)
+	
+	process
+	{
+		$script:exchangeschema
+	}
+}