+++ ForestManagement/functions/forestlevel/Get-FMForestLevel.ps1
@@ -1,0 +1,23 @@
+﻿function Get-FMForestLevel
+{
+<#
+	.SYNOPSIS
+		Returns the defined desired state if configured.
+	
+	.DESCRIPTION
+		Returns the defined desired state if configured.
+	
+	.EXAMPLE
+		PS C:\> Get-FMForestLevel
+	
+		Returns the defined desired state if configured.
+#>
+	[CmdletBinding()]
+	Param (
+	
+	)
+	process
+	{
+		$script:forestLevel
+	}
+}