+++ ForestManagement/internal/tepp/exchange.tepp.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'ForestManagement.ExchangeVersion' -ScriptBlock {
+	& (Get-Module ForestManagement) { (Get-ExchangeVersion).Binding }
+} -Global